requirements.txt
# Core Dependencies
gradio>=4.0.0
groq>=0.4.0
python-dotenv>=1.0.0

# PDF Export (Optional but recommended)
reportlab>=4.0.0

# Development Dependencies (move to requirements-dev.txt if needed)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
